Add files via upload
A python script that helps determine phenotypic resistance in TB Drug resistance line probe tests. You input the bands seen on the probe strip and it works out which TB drug resistance or sensitivity phenotype is present for that sample. Was contemplating adding a nice tkinter or qadesigner GUI to make it more visually appealing.